Map tokens from TTS responses to HTML
Added mapping between the tokens received from the TTS server and the
"words" in the html. Tokens are stored in the utterance elements and
are assigned a position attribute, which is the index of the start of
the corresponding html substring.

Cleaner received an overhaul to store information about the cleaned
tags. Previously, this information was just thrown away. The other
classes in the "PHP-chain" was updated to handle the new information.

Most test, both PHP and Javscript, was refactored to make them easier
to read.

Bug: T140105
Bug: T149174
Bug: T149799
Change-Id: Ie784328fa3d7bcf7941b6b89146687272fe3b0ca