Skip back sentence
Enabled skipping back to the previous sentence while playing. If the
current sentence has played for less than a certain number of seconds,
it's played from the beginning. This threshold can be set in the
configuration and is 3 seconds by default. Button and keyboard
shortcut (ctrl + left arrow by default) were added.

Bug: T133687
Change-Id: Ib63a013f525ce41f4e03bccf115ebe2a15fe4992