Floating control panel for player
Moved the buttons for controlling the player to a separate <div>, that
is placed at the bottom of the page. The control panel is fixed to
this position, even when scrolling. When debugging, the control panel
is placed above (Y-axis) the debug panel.

CSS linting was changed from csslint to stylelint.

Bug: T151887
Change-Id: I17ebf3ab3a32dba2b28715bc3053b6b2746bd4f3