Prevent keyboard shortcuts from triggering buttons.
keydown events with the same key combinations as keyboard shortcuts
don't fire. This is what triggers button click when a button is in
focus and space is pressed. This was implemented generally, i.e. not
only for space, since it's probably the preferred behaviour if there
are similar cases for other keys.

The default shortcuts was changed to use shift and alt instead of
control. This was done because there are some standard keyboard
combinations that were blocked, e.g. skipping word forward/backward
with ctrl+arrow when typing.

Bug: T146459
Change-Id: I8562ee4b78e50e29a35a4147b75b55bb3d6bc464