Add help button in player GUI
The button takes the user to a help page for Wikispeech. The page is
defined by the config variable `WikispeechHelpPage`, "Help:Wikispeech"
by default. The button only appears if the variable is set.

Changed button tests to use stubs instead of spies, as the function
for the help button changes redirected the browser.

Removed button ids that weren't needed.

Bug: T151886
Change-Id: Iba29fcb7f8102bfa3f82ca62794ed6ea4cff5921