Add API for segmenting text
The API will initially be used to clean and segment HTML. It can also
respond with only cleaned text or the original output. This can be
used for debugging or outside Wikispeech.

Cleaner and Segmenter is now instantiated, since there were a few
variables that were passed through multiple functions.

The config variable for removed tags now takes boolean or string as
values (and not objects), as it currently only allows CSS class to be
specified.

Bug: T164250
Change-Id: I78db6c6a64e9d04e6907d2eb96af47ca368cf82e