Use API to retrieve utterances
Utterances are now stored in a variable instead of HTML element. A
request is sent to the Wikispeech API action and the segments in the
response are stored. This is done on page load.

Default values for `removetags` and `segmentbreakingtags` are read
from the config.

Implemented specifying multiple CSS classes for tags, when listing
tags that should be removed.

Checking if the current page is in a namespace where Wikispeech is
activated was moved to Javscript code. It also uses id rather than
name.

Removed the logic for moving the player when the log panel was
added. This probably wasn't a good idea to begin with. If it turns out
that it's really needed, it should be implemented in some other way.

Moved logic for handling elements wrapping text content to back end.

Bug: T164251
Bug: T146454
Change-Id: I2bb7c2ae83b6aae2a1435e7041d8a5480e37f616