Buffer animation is hidden too early
With resepect to adding a buffer animation (loader)
when the audio is not available it was noticed that
there are instances when the loader is hidden before the audio starts playing.
This occurs mostly when several sentences are skipped in a row.

I order to solve this the following was done:

* The listener 'canplay' of the audio was removed from stopUtterance().
* A test was added to catch the same situation.

Bug: T175124
Change-Id: I745739594cb837a4a6b6ae5975df1a8dcf9919db