Recite selected text
If any text is selected, only that text is recited. Playback starts
with the first selected word and stops with the last. Partially
selected words are read whole.

Added a smaller selection player that appear when text is
selected. This player has a play/stop button, like the one in the main
player.

Since it proved quite tricky to combine reciting selections and
highlighting, highlighting is disabled when reciting selected text.

Audio elements are now created when utterances are created, rather
when the audio is loaded.

Reworked how request results are handled. Now, jqXHR is used to allow
adding callbacks to requests in progress. This also replaced the
utterance variable `waitingForResponse`.

Added ext.wikispeech.util for functions that are used in any module.

Fixed tokens' offsets, that were sometimes incorrectly calculated.

Fixed some things to improve the generated documentation. This will
continue in an upcoming patch that will refactor the code.

mw.loader is now used to load Wikispeech modules and to check that
everything has been loaded before setup.

Bug: T133680
Change-Id: I6fb92bff5947735105f62819017dd6732e7730bd