build: Remove unused 'qunit' from package.json
These dependencies are for the Node.js process used in 'npm test',
which in Wikispeech uses various Grunt plugins, but not qunit.

Also remove 'version' and 'license' fields, which are not
meant to be set if private:true, which indicates it isn't published
as a Node package, but only used for development/testing.

The actual package information is in extension.json, which already
has the license set.

Change-Id: I48a70fd76f798d044ca4ee1c2bc679bc4525f5b4