Wikispeech: Remove HtmlGenerator
This class was created before Javascript was used to request utterances.
It may become relevant again if and when we want to implement support for
non-Javascript usage. For now though, we remove since it's just dead code.

The first implementation stored utterances as HTML elements when rendering
the page. This is now handled with Ajax in the Javascript modules
(specifically Storage).

Bug: T244345
Change-Id: I974e2b18bc0ae490e306b28ae3e007e372cab84f