Wikispeech: Remove use of Utils.php
This was previously only used for the one class HtmlGenerator which was
removed in  T244344. Furthermore, the only method it has is was calling
private methods, which probably indicates that something is wrong with
the tests. If we really need such a utility, there is a MediaWiki class for it.

Bug: T244345
Change-Id: Idfede6272460b9bf8f27adc2205c2e847a2621fe