Schedulable cleanup job of expired utterance files in file backend
Iterates all files in utterance file backend and flush out those
that have a timestamp older than given time to live. This allows
for utterance metadata, utterane audio and synthesis metadata to
be out of sync but still be cleaned up.

But job is never scheduled. Investigations is required to find out
how heavy of a procedure this is on large distributed file backends
before we decide on when and how this should trigger automatically.

Includes maintenance script for manually triggering job.

Bug: T255001
Change-Id: I23b7e9717850a1277428219a41aecf1613433157