Connect API to utterance store
Introduces use of base64 audio rather than working with URLs.
Introduces the use of Utterance store rathar than accessing Speechoid directly.

Due to the current state of the API, which pass on text from the client
rather than working with segment hashes and page ids, this implementation
converts the text from client to hash and use page id 0 for all utterances.
This is a temporary solution that functions. It demonstrates the intended use
which will be implemented in a near future patch as the required changes are
added to the segmenting API and client end code.

Bug: T249203
Bug: T249200
Change-Id: Ib3082b5fe2e7e93739fcd75def39847849c9d328