Add option for showing player
Adds a tab in the action area (next to "Read", "Edit" etc.) called
"Listen" which loads modules and shows the player. The tab only
appears if it is possible to listen to the page and the user has
turned on Wikispeech in the settings. Adds a new user setting that
allows automatically loading the player on any valid page.

Bug: T243579
Change-Id: I4138bcaa31476ab9be6ccd92d03098d8c66f4e37