Add revision and segment id to listen API
If the parameters are supplied, the utterance cache will first be
checked. If a match is found it will be returned and only otherwise
will a request be sent to Speechoid.

Parameter "input" changed name to "text" to better specify what it
is. It is still possible to send requests with "text", but it's not
currently in use.

Bug: T248162
Change-Id: I81265230f59b798ebc250c9cb104e39ec0d30309