Allow listening to page even if edited
This ensures that a user who has started listening to a page can continue
listening despite the page having meanwhile been edited.

If the page is deleted listening will be interrupted.

If the revision is suppressed, listening will be interrupted, unless the user has
the permissions needed to view such pages.

If the page is edited before the page is segmented, listening will not commence.

Bug: T261753
Bug: T246084
Change-Id: I4550e3e955e0f7093e8c4874b66c33ef5c444099