Replace applicable @todo with @note in PHP-code
We have used @todo to highlight ideas with the code for
future consideration. This might confuse developers and
make them belive that they are requirements, when it
really is more about thoughts that may or may not improve
the stability of the system.

Patch also contains removal of @todos that are no longer
applicable, and minor typos in close proximity.

Bug: T263121
Change-Id: I148307e44414e0d026495a25c7a4e2474a609609