Hotfix: Maintenance script classloading
When introducing FlushUtterancesByExpirationDateOnFileFromFileBackendTest
we moved initantiating dependent classes from function scope to class fields
that were instatiated in the constructor. This caused the script to fail
when executing it from CLI as classloading occurs after constructor but
prior to the execution function.

This hotfix moves the instantiation of the class fields away from
the constructor to the exection function in order to make it work again.

Change-Id: I3ef046d632fa4e3be2a0b192dad1c7c993f23326