Throw exception rather than log error on non existing filebackend group
Previously we just logged an error if the filebackend group defined
in config was non existing in LocalSettings.php. This allowed us to
keep running without using the cache, synthesizing new audio for
each and every request. But the feature was never implemented so
it makes more sense to throw the exception early with a meaningful
message, compared to throwing a nonsensical null exception further
down the line.

Bug: T268664
Change-Id: Ia269647dbeea05c09e1f90357507ccdb20912e76