Lexicon edit communicate directly with Speechoid lexicon.
Stears all communication directly to Speechoid rather than
communicating through LexiconHandler. This means that we are
bypassing the local lexicon, which is OK since the current
implementation is a temporary transient solution.

We will go back to using LexiconHandler after implementing
the persistent local lexicon with revision and rollback features.

Bug: T279050
Change-Id: I3e588beca42170c2de1f4f78af2875f40275d75d