Optionally mark utterance files with TTL in Swift
To avoid invoking extra I/O operations on the utterance
file backend when flushing out files. In case of setting
WikispeechUtteranceUseSwiftFileBackendExpiring to true
while using a Swift backend, the extension will set a TTL
when creating the file and skip the delete operations on
flush. In order for this to work the administrator of the
Swift file backend must configure it to accept TTL on create.

This code will also require a future MW core patch that
accepts (doesn't filter out) X-Delete-After header passed
to the create file operation options.

Bug: T280244
Change-Id: If783fa7986c418be5853f540d2fdd41d2953ad3e