Read user options from user subpage on consumer wiki
Allows the user to set user options by adding a subpage to their user
page with the name "Wikispeech preferences". The page should contain a
JSON object with the options to override.

Also added template code to be used when adding as gadget or to user
scripts (common/global.js).

Bug: T281667
Change-Id: I4b8ce72c05304f715494378aea7f4bbab710be50