Log metrics from ApiWikispeechListen
Logs JSON to file set in WikispeechListenMetricsJournalFile.

Uses optimistic file locking on top of flock() without queuing.

Has support for archiving journals, but this is never invoked.
The idea is to implement a maintainence script that will execute
on the crontab, or by queuing a job now and then that checks if
the journal has grown past a certain size.

Bug: T286579
Change-Id: I4e09e12a8790e31446bd5b309bcb31ecf8c33e6a