Add option to edit preferred to Special:EditLexicon
If "preferred" is checked, `preferred` is set to true for the item. If
it isn't checked, `preferred` is removed from the item. This mimics
the behaviour of Speechoid and makes sure the lexicons don't get out
of sync.

If "preferred" is checked and there is another item in the same entry
that had `preferred` set to true, its `preferred` is removed.

Bug: T284060
Change-Id: I2586161c7c9b0760c704c996f3f56a3ca5444744