Properly segment titles with namespaces
A change since MW 1.35 is that the title element contains <span>
elements. If the title displays a namespace, there will be three
<span>s: one for the namespace, one for the colon and one for the page
name.

The segmenting was assuming that there was only one text node in the
title element, possibly wrapped in an element. Now all the text nodes
are added as separate segments.

Also:
* Removes related code dealing with empty items being added to
segments.

Bug: T322402
Bug: T255152
Change-Id: I51e1756ac1232b46ca3aa1bad1b4d5172cbeef46