Unbreak tests
- Bump required version of MediaWiki to 1.41
Needed for the namespaced Status class.

- Use namespaced MediaWiki\Status\Status
PHPUnit allows configuring a mock to return a value if the
types of the original method and the mocked value have an
identical name or the returned value is a subclass of the
type of the original method, which prevents using a mocked
value with an alias class.

- Add tests to the database group
Unless explicitly added the database is no longer available.

Change-Id: I6f01e199a234d4e2cf31f4c12e17f24c55f12f20