Redirect logged out users to login page if they lack right to edit lexicon
When a user is not logged in and they don't have the
wikispeech-edit-lexicon right they will be redirect to the login
page. After login they will be brought back to the lexicon editor with
the original URL parameters. If they still don't have the right
they'll get a permission error as usual.

They redirecting only happens if the new config variable
WikispeechEditLexiconAutoLogin is true.

Bug: T339840
Change-Id: I1ea7756cadcc0c8c8a28b251b21dbd65d4c2297f