Empty objects remain of type stdClass in LexiconEntryItem::properties
If an empty object is present in the structure of a JSON-object, then it is per default cast into an associative array.
We don't want this, since Pronlex expects certain types of LexiconEntryItem::properties. Therefore, we need to make sure that empty objects are of type stdClass.

Extra check on lexiconWikiStorage::saveLexiconEntryRevision() to remain lemma of type object

Bug: T279916
Change-Id: Ieb2f37be284d1779d219a176a6d2e4f64f4af28f