Read extra information for certain elements
Some elements, like links, lists and image, have extra information that the
user may want. One thing is just the fact that there is an element of a
certain type. There may also be element parameters that are relevant.

This introduces a framework to add that extra information. A basic
implementation for links is included as a proof of concept. Proper
implementations for it and other elements will be added in future patches.

The extra information is tracked in the segments by adding parts of content
(POC). This term refers to any elements in the content that has relevant
information beyond its text.

As a result of the extra text added to segments each segment that contains any
POC will be stored as a separate utterance. A new parameter is added to
wikispeech-segment, but isn't needed for wikispeech-listen because of this.

A user option is added for using POC. Due to technical complexities
highlighting words is disabled when this option is enabled.

Known issues:
* Doesn't announce links when selection starts at the beginning or within
   the link text (T402258).

Bug: T387739
Change-Id: Id8d0f38a742602bcceb7fa57277d86c302c3a339