Add a view link to edit lexicon entries as raw text
The link has "action=edit" and the extra parameter "raw=1". If it's set in
combination with "action=edit" the view for editing wikitext is shown rather
than redirecting to Special:EditLexicon.

The link only shows if the user has the new right
"wikispeech-edit-lexicon-raw". This right is not given to any group be
default since it should only be used if there is an error in a lexicon
entry. Editing raw JSON can easily break an entry if you don't know what
you're doing.

Some changes had to be made to tests for PlayerHooks because of how they ran
the tested functions.

Bug: T311047
Change-Id: I2762bcba83464a032c24675b32f6e9821a057d03