Remove `mw.wikispeech`
This was used to store variables that needed to be accessed be the different
components. These are now assigned in `Main`.

The remaining things in `mw.wikispeech` were `producerUrl` and `consumerMode`.
Both of them were used when running as a gadget. This adds a new config
variable for the producer URL. It's also used to check if the frontend is run
as a consumer so the producer mode variable is no longer needed.

Bug: T277733
Change-Id: Ic2d11fdaf8bbb2d8fa3624478ff918f063989396